Check for isdestroyed before setting state when we're closing a popper (#54)
I have an issue where when I make the call to close the panel and then 
transition away shortly after, the transition happens in the middle of the close
and when it gets to line 193 it errors out because the component has already
been destroyed. This should fix that issue.