Allow independent setting of font awesome weight
Previous to this commit the `fas` class would always be set and if
another weight was selected it was ignored.

This manifested when trying to include an information icon where with
the `fas` (font awesome strong) class the negative space is filled in to
be bolder and we were unable to use the `far` (font awesome regular)
class without this fix.

Co-authored-by: Chris Elten <chris.elten@addepar.com>